makefile: Change linker output format to ELF, add objcopy stage
Update the makefile to output the kernel as an ELF, then add an
objcopy invocation to convert the ELF to a binary file that can
be loaded by the bootloader.

Using an ELF has many advantages, especially for debugging since
symbol information can be viewed with `readelf --symbols` and
section load information can be viewed with `readelf -l`.